Parse plan to find needed columns and pass it down to AM.
This commit implements
- boolean where AM can specify it leverages column projection
- walk scan plan to find columns of interest

No AM currently leverages the passed in projection list but all the
info is available for it to leverage the same.

Revert "Add argument to scan_begin to pass columns to be projected."

This reverts commit 53a7c76ece34ea70557214b65043c1c8d7c3b9c4. I think
better to add new interface for begin_scan using column
projections. That avoids adding argument to all the other AMs which
don't support it. Irrespective anyways at Executor level need to know
if wish to pass down column projections down or not, so best can call
the API which is applicable.