Fix confusion on different kinds of slots in IndexOnlyScans.
We used the same slot, to store a tuple from the index, and to store a
tuple from the table. That's not OK, when the index contains fewer columns
than the table. It worked with the heap, because heapam_getnextslot() stores
a HeapTuple to the slot, and doesn't care how large the tts_values/nulls
arrays are. But with zedstore, which stores a virtual tuple, it caused
memory overruns.