Implement VACUUM, to remove aborted and deleted tuples.
The VACUUM uses the UNDO log to quickly find all the garbage. But we still
need to scan all the indexes.

This introduces the concept of DEAD tuples. In the long run, I want to find
a better solution for vacuum that doesn't leave behind DEAD tuples, but
this will do for now. I think we want to have DEAD tuples for other
purposes anyway, to "prune" items like in heap.