Trim the UNDO log more aggressively.
At the beginning of each scan (or to be precise, when the first visibility
check in a scan needs to be peformed), we scan the tail of the UNDO log
and try to advance the oldest undo recptr as far as we can. That's an
important optimization for the case that you load a table, and then
immediately query it in the next transaction. We used to do that
aggressively before the commit that introduced VACUUM support, but that
patch removed it. This commit brings it back.