Remove b-tree single-insert code, always use multi-insert.
Eliminates the code duplication. IMO it makes sense to think of
a single-row insert as a special case of a multi-insert, anyway. To
optimize single-row inserts in the future, I think we'll want to add some
kind of buffering, so that a stream of single-row inserts can be turned
into multi-inserts within a transaction. So that fits the multi-insert
interface pretty well.