Fix assertion failure in 'copydml' test.
If a deleting transaction aborts, and then the row is deleted again,
the code to check the visibility of the tuple afterwards failed an
assertion, because the "prev" undo pointer pointed to the aborted
deleting transaction. Aborted actions are not interesting to anyone,
so fix that by not keeping the "prev" undo pointer in such cases.