Implement UPDATE so that it cannot "run out of TIDs" on the page.
The new tuple version can now go to any page.

This is still incomplete, in a concurrent UPDATE/DELETE of the same tuple
can screw it up, because the old item isn't locked in any way, between
checking that it's updatable, and actually updating it. My plan is to
implement tuple-locking first, used also for FOR UPDATE, and use that here,
too.