zedstore: fix Alter table rewrite for add column.
At beginscan, zedstore, unlike heap, requires the tuple descriptor of
the tuple being requested. Previously, beginscan used the tuple
descriptor from the Relation. This did not work for zedstore in the
case of a table rewrite because the catalog is updated first, hence
tuple descriptor in relation object reflects new table format. The
tuple descriptor from the tuple tableslot reflects the tuple format of
the data we are about to scan.

So, to fix delay initialization of attributes until getnextslot, which
is where we have access to tuple tableslot.

Note: Ideally, now we can revert the changes made for alter table
alter column, but they are good for sanity check. So, keep for now.

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>