zedstore: fix alter table add column with default values.
To accomplish the stated objective following code has been added.

In the case of ALTER TABLE ADD COLUMN with a DEFAULT value, a btree is
not created for the column, as there is no data to store. The default
value is stored in catalog in pg_attribute.

Assumption was all the columns have data for all the tuples. This
doesn't hold true given above fact.

So, when scanning the table, it's problematic if we assume that we
will be able to determine the number of rows from any attribute in the
table. Thus, we have to ensure that the columns being projected from a
scan will include at least one column which has all the TIDs in table.
Which can be used to determine the expected number of rows for all of
the columns. Dropped columns are similarly invalid, as they are not
guaranteed to have the same number of rows as are active columns.

Function to fetch and fill missing attributes from the catalog
zsbt_fill_missing_attribute_value() is being added. This is called by
zedstore AM layer to fill in missing values for table. So, unlike
heap, for zedstore, executor doesn't need to fill these values.

btree scan callers use the tid for flow control.  btree scans set
isvaluemissing to indicate to the caller that it cannot rely on the
tid and now must use isvaluemissing flag for flow control.

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>