Implement Array items
An array item represents N consecutive TIDs and their datums, where each
element has the same visibility. Using array items makes the table much
smaller, when loaded with COPY, because the UNDO pointer, the TID, and
other header fields are not repeated for each item. Ideally, compression
would squeeze them out anyway, but in practice that doesn't seem to work
that well.

This also includes a bunch of changes and fixes to memory management,
handling toasted and unaligned datums, and ANALYZE. They're not all
directly related to array items, but I bumped into them while working on
array items, and it didn't seem worth the trouble to separate them out to
separate commits, since zedstore as whole is still so much work-in-progress.