Optimize the way datums are extracted from array items
Instead of walking the on-disk format array one element at a time, when
getnext is called, extract all the items into an array of Datums in one
go, and return elements from the Datum array one by one. Doing more work
in batches is more efficient.

Also, inline the fast-path code to return the next element from the
array.

Now that we're doing more work upfront, push checking for the "end TID" to
the btree scan code, to avoid doing that work when the end of a range has
been reached. zsbt_begin_scan() now takes an 'endtid' argument, and the
scan will stop when that TID is reached.