Fix alignment of stored datums.
Datums in array items are now aligned. The rest of the system requires
proper alignment, so it was either that, or make additional copies to
align the datums after pulling them from the array. In practice, that's
not a big deal, because "packed varlenas" with short 1-byte header are
never aligned, and anything less than 128 bytes uses the short varlenas.
So alignment is only needed for datums larger than that.

Also align the ZSBtreeItem structs in a compressed item. Otherwise, there's
no guarantee that the datum or array of datums stored in the item is
aligned either.