Zedstore now has first btree to store visibility information.
To simplify code and easily tackle with zero column table or handle cases with
add/drop columns, zedstore now has special btree at index 0 of meta-page which
is always present. This btree is always present and hence can be reliable used
for getting TIDs present in the table. Scanning code always adds this column to
projection list as first column to scan.

So, now this column becomes the special column based on which TIDs are assigned
instead of column 1 earlier.

This code currently just adds the new btree but doesn't yet eliminate storing
the undo info from other columns. That would be done as follow-up patch.