ALTER TABLE ADD COLUMN with no default works
If an attribute is not present in the table, previously, we validated
that it was because the attribute had atthasmissing true in the
catalog. That does not work in the simple case where the attribute was
added without a default value. In this case atthasmissing is
false. Now, simply return null in this case for the attribute.

Validation gets lost but seems need to trust the relation descriptor
over actual table file, as it has reality about the number of
attribute table has.

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>