Speculative insert works for zedstore
Previously, an insert with an on ON CONFLICT DO UPDATE would error out
with an error that zedstoream_insert_speculative() was not implemented
yet.  To implement it, when inserting an item in zsbt_multi_insert(),
set the speculative insert token in the item's undorecord.
Once the insert is completed, reset the speculative insert token in
place in the undorecord.

While completing the speculative insertion, if the insertion has to be
aborted because of a conflict, this abort scenario is not currently
handled. It is a TO-DO. So, error out in these cases.

When checking for uniqueness when attempting to insert into the index,
when doing the visibility check in zs_SatisfiesDirty(), the speculative
insert token from that item's underecord is used to set the speculative
insert token in the snapshot

Co-authored-by: Ashwin Agrawal <aagrawal@pivotal.io>