Leave out visibility information from attributes other than the meta attr.
It's more efficient, and it seems a bit strange to include an UNDO
pointer on attributes, when it's only the pointer on the meta-attribute
that matters.

The other attributes still have the capability of holding UNDO pointers,
although it's not currently used. That becomes significant when we start
doing in-place updates; an in-place update only needs to modify the changed
attributes, but those attributes will need to have a place for the UNDO
pointer for that to work.

Fix the scanning code to deal with that. It got confused, if the
meta-attribute showed that a row is invisible, but the other attributes
were visible. That case didn't occur before, because the same visibility
information was always copied to all the attributes.

Why did I do this right now? I was working on Free Page Map, and wanted to
allow deletion of dead items more aggressively at VACUUM, so that the
free'd space could be put on the Free Page Map and recycled. But that
exposed the issues with the scanning code, and one thing led to another.