Use TupleDescriptor from TupleTableSlot during scans.
TupleDescriptor from Relation object can't be trusted at AM layer, as
during ALTER TABLE rewrite, relation object passed to AM layer
reflects new TupleDescriptor and not old. Scanning zedstore needs to
know exact TupleDescriptor with which the data was written to
disk. Hence TupleDescriptor from TupleTableSlot needs to be leveraged
as that maps to what's on disk.

Also, as part of this removing the previous addition of storing attlen
and attbyval in root of meta-page. Its not needed and also anyways,
just, those two fields are not enough. Need more information from
TupleDescriptor like attmissing and also the actual missing values
stored in catalog for attributes with attmissing = true.