Fix bug in how 'undo_record_needed' flag was set.
If there is a chain of UNDO records, because a tuple was locked, we
must keep the undo-pointer on the item even if the oldest undo record
in the chain is no longer visible to anyone. Because the later records
still are. This was causing the 'nowait-2' isolation test to fail.