Remove the ZSBT_DELETED/UPDATED flags.
It was a bit difficult to keep them set correctly. For example, a DELETE
set the flag, but if the transaction later aborted, the item would still
be marked as DELETEd. We could fix that, but the flags are actually
redundant with the UNDO pointer. If the UNDO pointer points to a DELETE
record, then the row was deleted, and we don't need a flag for it.

This fixes the "could not lock updated tuple" and "could not lock deleted
tuple" errors, which were caused by the above scenario.