Alternate answer files for default_table_access_method=zedstore make check.
Zedstore doesn't use separate toast table. So, most of these tests
need separate answer file to pass make check for zedstore. We can
definitely rewrite the tests to check for toast in different way, like
have function which based on AM type performs check if toast has data,
which can leverage `pg_zs_page_type()` to find toasted zedstore pages
and return true or false. That way still get the required coverage for
the zedstore as well. But currently just lazy to do it and hence these
alternate answer files.

So, answer files differ due to phyiscal tlist optimization is disabled
for zedstore, and hence not all the columns of the table are
projected.