Implement relation_copy_for_cluster callback, for CLUSTER.
There's no support for CLUSTER by sorting yet. Currently, it always scans
the index (unless it was a VACUUM FULL, with no ordering)

This adds new zsbt_reset_scan() function, to reuse an open ZSBtreeScan
to fetch a different TID. Essentially, to change the 'nexttid' of the
scan correctly on the fly. We probably should use that for index fetches
too, but that's not done in this commit.

I had to add an alternative expected output file for the 'cluster' test to
make it pass with zedstore, because it looks at whether some tables have a
TOAST table, which zedstore tables don't have.