Avoid creating blocks for empty table on index build.
create table foo(a int, unique(a)) was creating meta-page and root blocks for
zedstore. So, after this operation table contained 4 blocks, ideally it should
have zero. This was causing many test to fail, due to plan differences between
heap and zedstore.

The problem was zsbt_descend() always passed forupdate=true to
zsmeta_get_root_for_attribute(). This caused blocks to get allocated even for
read-only operation.

Now, modified zsbt_descend() to take read-only argument and pass it down to
zsmeta_get_root_for_attribute(). Hence, for the mentioned case now, it will
return with no rows and not create blocks for the table. Only first insert to
the table creates the blocks for empty table.