Split zedstore_btree.c into three files.
zedstore_tidpage.c deals with the TID or META tree, which contains no data,
  but just the visibility information (UNDO pointers) for each row.
zedstore_attpage.c deal with the actual attribute trees, containing the
  user data
zedstore_btree.c contains common code between the first two. Notably, it
  contains the functions to deal with internal pages.

There were a few reasons for this refactoring:

1. zedstore_btree.c had grown quite large. It's nicer to work with smaller
   files.
2. I think we'll need to rewrite the leaf page format for performance
   reasons. Attribute pages don't really need the UNDO pointers, and we'd
   expect to compress them, and they're basically append-only as all new
   rows are assigned TIDs in increasing order. And the TID pages don't need
   the data, and we possibly don't want to compress them, but need to be
   able to update existing items  on DELETE/UPDATE.

The on-disk format is unchanged in this commit, but I plan to work on that
(see reason 2 above).