Fix code to step to next btree page.
Without this, we might get into an infinite loop if we encounter an empty
page. We step to the next page, but because we hadn't updated 'nexttid',
we decide that the next page is not actually the page we're looking for,
and descend from root instead. It will descend back to the original page,
and try to step right, and so on.