speculative insert abort works
If a transaction inserts a tuple speculatively into the table and then
finds a conflict when attempting to insert into the index specified in
the ON CONFLICT clause, it marks the speculatively inserted tuple
in the table as dead and does the specified ON CONFLICT...DO action.

In addition to marking the tuple as dead in the meta column, the actual
items must be removed from the attribute btrees. This is because once
the tid is marked as dead, it is valid to be reused, and, without
removing the btree items, a zombie value will be returned once the
tid is reused.

Also updated update_indexes according to if the column being updated is
an index -- this is the same as what heap does.