Use slightly different item format for attribute and TID trees.
Attribute items don't have a field for UNDO pointer anymore. It's not
needed, since visibility is determined by the TID items. (I think we'll
need the UNDO pointer on attribute items again if/when we implement
in-place UPDATEs in the future, but let's cross that bridge when we get
there.)

TID items are no longer compressed. That obviously uses more space, but
the TID items are very compact anyway, because a single "array" item can
cover a range of TIDs. We might want to add that back later, but probably
an even better approach for the TIDs is to invent different item formats,
with a dead-item bitmap, varbyte encoding etc. Using a compact item format
instead of compression makes updates/deletes faster, as we don't need to
decompress and recompress so much.