Add more zedstore specific answer files.
The answer file difference for triggers test should be looked more
deeper. Like why the failure is happening in delete for zedstore vs
hapens in get tuple for trigger for heap. But for now its okay to have
answer file for zedstore.

rowsecurity test needs answer file just because analyze for zedstore
is not able to compute common_vals for the column.