Cache tree roots and rightmost leafs in backend-private memory.
This considerably speeds searching and insertions, as you don't need to
hit the metapage all the time, and you also don't need to descend the
tree when inserting to the rightmost page (almost all insertions go to the
rightmost page).

TODO: We should propose the caching mechanism on pgsql-hackers. I just
added a new field to SmgrRelationData, beside smgr_targblock et al, but
I think we should move targblock and the cache FSM and VM sizes to an
AM specific struct, too. Or perhaps we should use RelationData->rd_amcache
for this instead of storing it in SmgrRelationData; not sure.