Simplify the FPM.
Instead of another btree, indexed by block number, use a simple linked list
to store the free pages. I was seeing self-deadlocks with the tree FPM, when
deleting a page caused an FPM page to become empty and also removed, which
required adding the page to the FPM again. We'll probably want to implement
something more sophisticated later, to avoid fragmentation and to avoid
having to dirty pages that are deleted, but not a priority right now.