Change the way VACUUM and marking items dead works.
The main change is that if a row is determined to be dead, when scanning
the UNDO log, it's immediately marked as dead in the TID tree. That allows
the UNDO log to be trimmed, even when there are aborted transactions or
committed deletions. A downside of this is that VACUUM now needs to scan
the whole TID tree to determine which rows are dead, not just the UNDO log.
It's probably not too bad, given that the TID tree is only a fraction of
the whole relation file, but we could optimize it further by keeping
something the freeze map. Perhaps flag pages that don't contain any dead
items in the upper levels of the TID tree, so that they can be skipped by
vacuum.

One reason for this change is that the EDB UNDO log patches assume that
the UNDO log can be trimmed (aka discarded) aggressively, whenever a
transaction becomes visible to everyone. Aside from that, it seems simpler.