Store NULLs with per-TID flags in attribute trees.
There used to be just a single NULL-flag on an attribute item. Now there's
a NULL bitmap. This allows storing NULLs and non-NULLs in the same array
item, which packs much more tightly when there's a mix of NULLs and
non-NULLs. On the flipside, straight runs of NULLs or non-NULLs use
slightly more space now, now that that there's always a NULL bitmap, but
seems better to avoid the pathological case. We might re-introduce a
per item flag later, if we want to optimize the common case of no NULLs
in the future.