Alignment fixes.
This alignment stuff is really tricky. I've fixed issues in this area a
couple of times already. This time, I chased a compiler warning about an
uninitialized variable, which led to these fixes.

TODO: We should reconsider how to pack the items in an attribute array.
Perhaps it would stop using the normal layout of varlena Datums, and
e.g. have a separate offsets/lengths array.

alignment fixes