Perform splits more evenly in the TID tree.
Instead of stuffing the left half as full as possible, aim for 50/50
splits, or 90/10 on the rightmost page. That avoids bloating the TID tree
on unfortunate update patterns, while still packing items fairly tightly
in the common case that the table is never updated.

This actually makes the cases that I've been testing with in
src/test/storageperf somewhat worse, but as we make the TID tree packing
smarter, this becomes important.