Pack items in TID tree differently.
This new packing scheme is more efficient when there are some deletes
or gaps in the TIDs. This prevents the TID tree from exploding when you
e.g. delete every other row.

TODO: It's a bit slow to work with. Experiment with different variants.

With this, the items stored on a TID leaf page follow the normal page
layout, with line pointers, again. So this reverts my earlier commit
6be3aab38f949a12b9cc32b8f27833c8a860b270 which optimized them away. The
items are now quite large, so the line pointer overhead is not that
significant, and it's easier to work with variable-sized items in this
new format, than try to force them all to be the same size.