Cache last delete undo record.
With this patch, if you perform a DELETE that deletes multiple tuples in
the same command, we can create a single UNDO record to represent a bunch
of deletions. That's a good thing, because the packing we use in the TID
tree is more efficient, when tuples share the same UNDO pointer. This
caching makes that happen much more often.