Fix rolling back a multi-insert.
We used to only mark the first item as dead, when undoing and insertion.
In a multi-insert, need to mark all of them dead. Otherwise, after
rollback, the non-first tuples are not rolled back.

While we're at it, clean up the UNDO record types a bit. Remove the 'tid'
field from the common record header, because half of the record types were
not really using it. In an insert record, it was used for the first 'tid',
but since all operations need to take the end of the range into account,
too, it's just confusing to have the first tid in the common header. Also,
change the meaning of 'endtid' to be the last tid + 1. That's a more common
convention throughout the code. And move speculative insertion token from
the common header to ZSUndoRec_Insert, too.