Fix vacuum.
In the commit to rewrite ZSTidArrayItem format, I originally used slot
number 0 to mark dead items, but changed it later to 1, and created
constant a #define for it. But I forgot to change this hard-coded
reference to 0. This caused the 'groupingtest' test to fail sometimes,
because VACUUM marked wrong rows as dead. (I'm a bit surprised it didn't
cause more failures, but oh well..)