Fix scanning with SnapshotDirty.
The visibility check when using SnapshotDirty returns xmin, xmax and
speculativeToken of the tuple to the caller, in the SnapshotData struct
itself. I missed that hack when I rewrote the ZSTidArrayItem format to
use the undo slots on each item. Rearrange things so that those fields
are again returned correcrly back to the caller from zsbt_tid_scan_next().