Change the ZSTidArrayItem format again
In the previous format, the UNDO slot number and TID of each tuple formed
an integer, and those integers were Simple8-b encoded into codewords. In
hindsight, it seems more intuitive to pack the UNDO slot numbers
separately. So there are now three separate parts in each TID item:
the TID list, as Simple-8b encoded codewords, the UNDO slots, and the
an UNDO slot number for each tuple, packed in 64-bit words, with 2 bits per
tuple.

Another reason for doing this is that I'm thinking that we might want to
include a TID list like this in ZSAttributeArrayItems, too, so that if you
remove a single TID in the middle of an array, we don't need to immediately
split the array into two. I.e. change it so that the TIDs packed in a single
attribute array wouldn't need to be strictly consecutive. I think that
might simplify things, and make the behavior more smooth when you delete or
update a few tuples. If we do that, the ZSAttributeArrayItems would store
just a TID list without UNDO slot numbers. I'm not sure how much code they
can share with ZSTidArrayItems, but keeping the two more similar is surely
better. Makes copy-pasting between them easier if nothing else.