Avoid "could not find UNDO record" errors if undo log is trimmed concurrently.
I sometimes got "could not find UNDO record" errors when running the
regression tests in parallel. I didn't really debug it deeply, but I'm
pretty sure that it was caused by two scans happening on a zedstore table
concurrently, probably on one of the "standard" tables like 'onek'. We
trim the undo log quite aggressively, even on selects.

What happened was that one backend determined that a tuple was recent
enough that we needed to fetch its UNDO record, to determine if it was
visible. So it called zsundo_fetch() on the undo pointer. However, another
backend trimmed the undo log concurrently, removing the block where the
undo record was. The record was in fact already visible to everyone in
that case, but the first backend was still working with a stale
recent_oldest_undo value. That would be harmless and expected in the
current zedstore design, except that zsundo_fetch() threw an error if it
couldn't find the record.

Fix by making zsundo_fetch() return NULL if the page the record was
supposed to be on was trimmed away. Change the callers to deal with that,
and update their recent_oldest_undo value.