Add alternate answer file for join_hash.
The failure for zedstrore is just because of the reltuples and
relpages numbers being off because analyze is sub-optimal. Planner
calls initial_cost_hashjoin which calls ExecChooseHashTableSize this
uses the stats to determine how many batches are needed. For one of
the queries in the test file, for example, the table "simple" is
created.

Fixing this will require fixing analyze.