Use Bitmapset instead of bool array for used columns as it's efficient
After proposing upstream a patch to get used columns during planning
which composed the used columns in a bitmapset, it made the most sense
to change the implementation in zedstore patch to use bitmapset as
well.

In the future, once the patch for extracting used columns at plan-time
makes it into Postgres, it should be a simple task to grab them from
the RangeTblEntry for those callers to whom it applies.

See thread here for reference:
https://www.postgresql.org/message-id/flat/CAAKRu_Yj%3DQ_ZxiGX%2BpgstNWMbUJApEJX-imvAEwryCk5SLUebg%40mail.gmail.com

Co-authored-by: Melanie Plageman <mplageman@pivotal.io>