Scan only required columns during parallel scans
Previously, for plans with parallel sequential scans, all the columns were
scaned for zedstore. This leads to unnecessary page reads and makes it
unefficient for column store. In this patch, the projection list required by
the plan is used in table_beginscan_parallel() to scan only minimal columns.

With this the runtime for zedstore parallel scans are faster than zedstore
non-parallel scans. Also, much faster compared to heap parallel scans.

Co-Authored-By: Ashwin Agrawal <aagrawal@pivotal.io>