Inlining these initialization functions seems excessive.
The point of inlining is to have a fast-path for the common case that the
initialization has been done. No point in trying to optimize away the
single call instruction in the rare case that initialization is actually
needed. Add an if() guard to the only caller that didn't already have it,
to have the fast-path, and remove 'inline'.
to the one caller that was missing it, and remove 'inline'.

Ok I lied: not all the callers of zs_initialize_proj_attributes_extended()
still have an if()-guard, but those callers are doing much more heavy
stuff, like descending the TID tree, that the extra call instruction won't
make any difference.