Refactor code to find and lock correct leaf page for TID.
Move it to a helper function, shared by TID and attribute tree scan code.
There's more similar code in them, but this one at least seems like a clear
case, that belongs well in zedstore_btree.c, becaause it only deals with
page level stuff and header fields, rather then the details of the items
on a page.