Free memory allocated for various buffers at end of scans.
This was causing a massive leak in ANALYZE, bitmap scans and sample scans,
which repeatedly open and close scans. We should modify those to keep
the scans open, the TID and attribute scans can be repositioned easily
without ending and restarting, but let's plug the leaks in any case.

I can still see a small leak in ANALYZE, but I'm not sure where it's
coming from.