Refactor the "state" logic in zedstoream_getnextslot().
This seems simpler.

Allow a TID scan reset to change the start and end of the range to scan,
in addition to the current position. This way, we don't need to close
and reopen the scan whenever we switch to a new range in a parallel scan.
The underlying TID scanning code can handle that fine nowadays, so little
reason not to allow that. It's faster, but also simpler.