Don't update metapage every time the undo counter advances.
There's a pointer to the latest UNDO page in the metapage. To get the
"latest" UNDO rec ptr, you now have to read the latest UNDO page, and
look at the last_undorecptr on that page. The only place where we need
the latest undo pointer, is when we're inserting a new UNDO record,
and we'll need that page to insert a new record anyway, so that's OK.

One advantage of this is that inserting a new UNDO record no longer
requires an exclusive lock on the metapage, which should improve
concurrency.

If there's no UNDO log at all, then you can't look at the latest UNDO
page. So we still keep an UNDO pointer value in the metapage, which
is the pointer at the beginning of the latest UNDO page, but it's also
set when there are no UNDO pages. In that case, it's the pointer to
the beginning of the latest UNDO page, if it existed. We only really
need the value, when there is no UNDO log, but we keep it on the metapage
as a cross-check in all cases.