Move zstid routines to separate header file.
This is so that we can use 'zstid' in client code. Why would we want to do
that? Because pg_waldump is a client program, and it links with
zedstoredesc.c, and needs to read the WAL record structs.

We don't currently use zstid in any of the WAL record structs, but that
will become necessary soon.