Fix bug in vacuuming away dead TIDs.
The loop to decode the old slot numbers from the TId item was broken.
This was causing some regression failures in the standard regression suite,
but curiously only if run in serial. So "make intallcheck" had more failures
than "make installcheck-parallel". I think that's because in parallel,
old tuples could not be vacuumed away so aggressively. I'm not sure why this
didn't come up before, or why I didn't notice the failures before. The
test that was failing was 'limit', but you could see the problem earlier
already, if you issued the query "select * from onek where unique1=69" in
between tests. The problem happened in the 'misc' test, where it deleted
all rows and reloaded them.