Optimize index scans.
No need to restart the TID and attribute tree scans from scratch for
every row. If we're lucky, the next index probe is on one of the TIDs
we already have decoded in the scan's arrays. That speeds up index scans
considerably, if you scan TIDs that are close to each other, which is
pretty common in practice.