Fix BRIN bug.
NOTE: I think this is an upstream bug. If the union function called
by FunctionCall2Coll call above the changed code returns the original
'newval', rather than a newly palloc'd value, we will store it in
column>-bv_values[INCLUSION_UNION]. But earlier in the function, if the
input is NULL, we make a copy before storing 'newval' in
column->bv_values[INCLUSION_UNION]. I think a copy is needed here too,
to make sure the value is in a long-lived memory context.

Likewise, if the function returns the other argument, we mustn't pfree()
it and then make copy. That's a reference-after free.

TODO: Post to pgsql-hackers. But fix on 'zedstore' branch now, so that it
doesn't block us.