Refactor attstream decoding to use less memory.
Introduce the concept of a "decoder", so that all the attstream
manipulation routines that need to decoder an attstream don't need to
set up a fake "scan" object.

The decoder will now decoder an attstream in smaller chunks. The old
decoder_attstream() function used to decode the whole attstream in one
go, but that can be expensive. I ran into out-of-memory errors on my
laptop, when I tested with a 50-column table full of NULLs. It compressed
so well that each page contained millions of elements, and we allocated
huge arrays to hold them. Now they will be decoded in smaller chunks,
requiring much less memory.

This probably makes random access to attstreams, i.e. index scans,
somewhat faster, too. But I haven't measured that.