Add attribute buffering, to speed up INSERT+SELECT.
When a tuple is inserted, only insert the TID into the TID tree on the
spot. The attribute data is buffered in in-memory buffers, instead. Only
when enough attribute data has been buffered, or the transaction commits,
write out the data to the attribute trees.

Speeds up INSERTs considerably, and reduces WAL volume, when you do insert
multiple rows in a transaction witout using COPY.