Implement "TID reservation", to speed up INSERT+SELECT.
This improves INSERT+SELECT performance further. When a row is inserted,
insert immediately 100 TIDs to the TID tree, and remember those in backend
private memory. When subsequent rows are inserted in the same
(sub)transaction, we just hand out one of those 100 TIDs, and don't have to
do any on-disk changes, until all 100 TIDs have been consumed.

If there are any unused reserved TIDs left at end of transaction, remove
them from the TID tree before finishing the transaction commit.

This works, because as long as the transaction is still in-progress, no
one can see the inserted rows, and hence won't care that we actually
inserted some extra ones. As long as we kill them before the end of the
transaction.