Fix deadlock if two backends try to extend the UNDO log at the same time.
Backend A:

1. Locks metapage.
2. Locks UNDO tail page
3. Sees that it's full. Unlocks metapage
4. Get new page from Free Page Map
5. Re-locks metapage.

Backend B:
1. Locks metapage
2. Locks UNDO tail page.

Because we keep the FreePage Map as a list, with the list head in the
metapage, getting a new page from the Free Page Map acquires the lock on
the metapage. That causes a deadlock risk: At step 4, backend A is holding
lock on the UNDO tail page, and acquires lock on the metapage. At the same
time, in step 2 in backend B is holding the lock on the metapage, and tries
to lock the tail page. Deadlock.

I ran into this, when testing tuple locking with pgbench. I had 20
concurrent sessions, each doing "SELECT FOR SHARE" on a small number of
rows.