Free attbuffers when deleting a tuplebuffer
Previously we did not free the memory allocated for the attbuffers when
deleting a tuplebuffer. This caused a memory leak when flushing tuplebuffers
multiple times within a transaction.

Ideally we would prefer to keep the attbuffers arround so that we can re-use
them, but we will need to think about the implications. For now this should fix
the memory leak.

This is spotted by Ashutosh Sharma.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>