Compress toast datums
Previously toast datums are stored in separate toast pages
uncompressed, which could be a waste of disk space and IO. This patch
adds compression to the toast datum using toast_compress_datum().
After the compression, depending on the toast datum size, either store
the datum in-line or in out-of-line toast pages.

Co-authored-by: Taylor Vesely <tvesely@pivotal.io>