Attribute-level FPMs and rel extension batching
To reduce the non-contuguity of blocks for attribute and tid tree
(especially leaves), which in turn improves I/O readahead and I/O
performance, we introduced the following:

1. Attribute level FPM
2. A new reloption: zedstore_rel_extension_factor, which designates the
number of blocks that will be fetched from the storage manager when we
request a single block. Out of the blocks fetched from the storage
manager, 1 block will be returned (to meet the request for a single
block) and the rest will be prepended to the attribute-level FPM.

From now on, the table level FPM, in ZSMetaPageOpaque, will only serve
UNDO pages.

Even TOAST pages, given their access pattern of being accessed
immediately following an attribute leaf, will be served from the
attribute-level FPM.

By increasing zedstore_rel_extension_factor (such as 16, 32, 512, 4096
etc.) in our experiments, we found that we get a very significant
reduction in I/O time for reads (as reported by track_io_timing and
explain (analyze, verbose, timing, buffers)) on rotational disks and
especially where the data was loaded into the same table by many
concurrent sessions (such as 16).

Notes:

1. We also use the term attribute-level FPM when we refer to the tid
tree (attno = 0).
2. Since we store the FPM heads for each attribute in the meta-page, we
hit the limit on the number of columns we can support. This causes the extra
wide table test to fail in create_table.sql and sanity_check.out. Added
FIXMEs for them to be done when we tackle meta-page overflow.

Discussion (including benchmarks):

Co-authored-by: Soumyadeep Chakraborty <sochakraborty@pivotal.io>